test(components): add generic unit test for hostHiddenStyles | bh
