+++ packages/components/src/styles/hover-media-query.spec.ts
@@ -2,7 +2,7 @@ import { hoverMediaQuery } from './hover-media-query';
 import * as jssUtils from '../utils/jss';
 import type { TagName } from '@porsche-design-system/shared';
 import { getComponentMeta, TAG_NAMES } from '@porsche-design-system/shared';
-import { addParentAndSetRequiredProps, componentFactory } from '../test-utils';
+import { addParentAndSetRequiredProps, componentFactory, getCssObject } from '../test-utils';
 
 const originalEnv = process.env;
 const style = {
@@ -70,19 +70,7 @@ it.each<TagName>(tagNamesWithJss)(
     expect(spy).toBeCalledTimes(1);
 
     if (type === 'return') {
-      // useful for debugging
-      // const mediaQueriesAndSelectors = Array.from(cssString.matchAll(/(.+) {/g)).map(([, selector]) => selector);
-      // console.log(mediaQueriesAndSelectors);
-
-      const jsonCssString = cssString
-        .replace(/"/g, "'") // replace double quotes with single quotes
-        .replace(/(.+) {/g, '"$1": {') // wrap selectors in double quotes
-        .replace(/ ([\w-:]+): /g, '"$1": ') // wrap css properties in double quotes, initial space is to skip media query values
-        .replace(/: (.+);/g, ': "$1",') // wrap css values in double quotes and convert semi colon to colon
-        .replace(/,(\s+})/g, '$1') // remove comma of last value
-        .replace(/}\n([^}])/g, '},\n$1'); // add comma after closing bracket if not nested
-
-      const cssObject = JSON.parse(`{${jsonCssString}}`);
+      const cssObject = getCssObject(cssString);
 
       Object.entries(cssObject).forEach(([key, value]) => {
         // potential media query
