+++ packages/components/src/test-utils/index.ts
@@ -1,3 +1,4 @@
 export * from './addParentAndSetRequiredProps';
 export * from './componentFactory';
+export * from './getCssObject';
 export * from './tag-names-constructor-map';
